requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
openai==1.58.1
python-docx==1.1.2
pydantic==2.10.3
python-multipart==0.0.20
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
playwright==1.49.1
pytest-playwright==0.6.2
